Upgrading the default ingenic_videocap and v4l2rtspserver pinned code
Newer version of ingenic_videocap copies audio frames to alsaloop.
Newer version of v4l2rtspserver has MP3 encoding support.

This is a customized version of v4l2rtspserver and uses the SHA hash
for the openmiko branch. A PR has been submitted but is pending so
for now use our own branch.