Fix a bug where some people flashing might get "tar: invalid tar magic"
On occasion the flash will work but because uboot doesn't totally erase
all the other partitions we end up in a situation where the jffs filesystem
is in a bad/corrupt state.

To fix this we can remove the extra jffs files. The real fix is to clear
out the rest of the filesystems on the flash chip with zeros but this
seems to work as well.

This commit also has busybox link more utilities up so we have access
to find and xargs commands (among others).