Work around for an issue where the rootfs.tar.xz file is corrupt
On occasion the rootfs.tar.xz file will be corrupted. Not entirely
sure of the cause. In addition the uboot flash process *should*
clear out the jffs partition but it does not. Frequently there are
files left over. This surfaces as the error:

tar: invalid tar magic

To get around this delete files in the /jffsroot on init if we
encounter the issue. This seems to stir the bits around and a
subsequent flash seems to work.