Fix for sometimes flashing and unable to mount rootfs
Clear out unused rootfs blocks on reboot.

The rootfs, appfs, drivers Wyze partitions have a total size of 8978432.
This is the most I can write with a stock u-boot bootloader.
However the mtd partition is 13828096 bytes and runs into the old config partition
This causes JFFS mount issues because when JFFS tries to mount it reads some of
the old stuff and corrupts things.

To fix this we will zero out the unused blocks. After they are zeroed
we will mount the jffs filesystem which will autocreate CLEANMARKER
blocks.