Use a custom logrotate package
This patches the logrotate binary so that it ignores permissions
and rotates anyway. This is because we use a FAT32 system
for the sdcard and it may not support unix style permissions.

This is a problem because the sdcard is used to get data off of
the camera when wifi isn't working.